alx_travel_app/alx_travel_app/requirement.txt
django
djangorestframework
django-cors-headers
drf-yasg
celery
rabbitmq

